Revert "Jshint fixes"
