--- .travis.yml
@@ -1,9 +1,0 @@
-
-language: node_js
-node_js:
-  - "0.10"
-  - "0.12"
-  - "stable"
-before_script:
-  - npm install -g gulp
-script: gulp
