--- js/asciiRPG/sprite.js
@@ -1,13 +1,0 @@
-
-var Sprite = function(){
-	this.image = "<   >";
-	this.map = "## ##";
-};
-
-Sprite.prototype.getImage = function(){
-	return this.image;
-};
-
-Sprite.prototype.getMap = function(){
-	return this.map;
-};
