--- package.json
@@ -1,21 +1,0 @@
-{
-  "name": "asciiRPG",
-  "description": "Text rendered 2d tile based game engine",
-  "dependencies": {
-    "del": "x.x.x",
-    "gulp": "3.9.x",
-    "gulp-jshint": "2.x.x",
-    "gulp-concat": "x.x.x",
-    "gulp-rename": "x.x.x",
-    "gulp-uglify": "x.x.x",
-    "merge-stream": "x.x.x"
-  },
-  "engine": "node 0.12.7",
-  "scripts": {
-    "test": "gulp lint"
-  },
-  "devDependencies": {
-    "gulp-jshint": "^2.0.0",
-    "jshint": "^2.8.0"
-  }
-}
