+++ tpls/player.txt
@@ -1,0 +1,20 @@
+   ___       ___       ___   
+  /mmm\     /mmm\     /mmm\  
+  \@,@/     \@,@/     \-.-/  
+   Omm'    q mmm p    'mmO   
+   ` n       n n       n `   
+   ___       ___       ___   
+ _/mmm\    _/mmm\    _/mmm\  
+ |6\\)/    |6\\)/    |-\\)/  
+  `mOn      `mOn      `mOn   
+  \   b      dd       \   b  
+   ___       ___       ___   
+  /mmm\_    /mmm\_    /mmm\_ 
+  \(//6|    \(//6|    \(//-| 
+   nOm`      nOm`      nOm`  
+   d   /      bb       d   / 
+   ___       ___       ___   
+  /-_-\     /-_-\     /-_-\  
+  \mmm/     \mmm/     \mmm/  
+  dmOm       mOm       mOmb  
+   ` ,       , ,       , `
