+++ tpls/title.txt
@@ -1,0 +1,30 @@
+                                                                                
+                                                                                
+                                                                                
+                                            __                                  
+         _.'---'''-.,        _..          _-  \            ____                 
+        \      _     |      |   \  /`-.  /  _-`__  _,     |    |    --_         
+         \    \  \   /  __   |   V`  _/`\_-`  |  \/  \     |    \  /  /         
+           \   \ /,./,-`  `-,|     <' /` _`-  /      \ _--_|     |/  /          
+            \    \  /   \  - \      `-| (/ _ |        /    \        /           
+             \   \  |    ``  |  |\    `-___-`|  /\/\ |  (_- | |    /            
+              \   \  \_    _/|__|   \_  |   /   /  \  \_  _/  |    /            
+               \,.-'   `--`           ``-    ``-    `--``` `--|   /             
+                                            +,_              `--_/              
+              ____   _____   _   _         /=-.` _   _   ____   _   _           
+             /    | /  ___| | | | |       //    | | / / |  __| | | | |          
+            /  _  | | (_    | |_| |      //     | |/ /  | |__  | | | |          
+           /  /_| |  \_ -_  |     |   __/,'     |   |   |  __|  \ V /           
+          /   ___ |  __)  \ |  _  | ,' ``:      | |\ \  | |__    | |            
+         /   /  | | |     | | | | | |     |     | | \ \ |    |   | |            
+        /___/   |_| |____/  |_| |_| `._,,'      |_|  \_||____|   |_|            
+                                                                                
+                                                                                
+                                                                                
+                                                                                
+                                                                                
+                                                                                
+                                                                                
+                                                                                
+                                                                                
+
