+++ workspace/simon/rpgTest.html
@@ -1,0 +1,121 @@
+<html>
+<head>
+<title> Pokemon Ash Key</title>
+<style>
+	body {
+		background-color: rgb(218, 7, 7);
+		font-family:Verdana,Arial;
+	}
+	#title {
+		width: 250px;
+		float:left;
+	}
+	#title h1 {
+		text-align:center;
+	}
+	#main {
+		display:block;
+		display:relative; 
+		margin:auto;
+		height: 90%;
+	}	
+	#getInvolved {
+		background-color:beige;
+		border-radius:2px;
+		padding:5px;
+	}
+	.clearfix {
+		clear:both;
+	}
+</style>
+<script src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
+<script src="asciiRPG.js"></script>
+<script>
+	var game;
+	var playerString;
+	
+	$(function(){
+		$.get("player.txt", function(response){
+			playerString = response;
+		}).then(function(){
+	
+			game = new World(document.getElementById("main"));
+			var terrain = generateTerrain(80, 30, 44);
+			
+			
+			var box = new GameObject();
+			box.sprite.image = generateBox(10, 5);
+			box.sprite.map = generateBox(10, 5);
+			box.x = 20;
+			box.y = 10;
+			game.room.add(box);
+			
+			box = new GameObject();
+			box.sprite.image = generateBox(30, 5);
+			box.sprite.map = generateBox(30, 5);
+			box.x = -20;
+			box.y = -10;
+			box.width = 30;
+			box.height = 5;
+			game.room.add(box);
+			
+			for(var i=1;i<=10;i++){
+				for(var j=1;j<=9;j++){
+					var tile = new GameObject();
+					switch((i+j)%4){
+						case UP:
+							tile.sprite.image = generateTile("^");
+							tile.mayExitDirections = [UP, LEFT, RIGHT];
+							tile.mayEnterDirections = [UP, LEFT, RIGHT];
+							break;
+						case DOWN:
+							tile.sprite.image = generateTile("v");
+							tile.mayExitDirections = [DOWN, LEFT, RIGHT];
+							tile.mayEnterDirections = [DOWN, LEFT, RIGHT];
+							break;
+						case LEFT:
+							tile.sprite.image = generateTile("<");
+							tile.mayExitDirections = [LEFT, UP, DOWN];
+							tile.mayEnterDirections = [LEFT, UP, DOWN];
+							break;
+						case RIGHT:
+							tile.sprite.image = generateTile(">");
+							tile.mayExitDirections = [RIGHT, UP, DOWN];
+							tile.mayEnterDirections = [RIGHT, UP, DOWN];
+							break;
+					}
+					tile.sprite.map = generateTile("#");
+					tile.x = i*10;
+					tile.y = j*5;
+					game.room.add(tile);
+				}
+			}
+		
+			game.room.add(game.player);
+			game.run();
+		});
+	});
+</script>
+</head>
+<body>
+	<div id="title">
+		<h1>Pokemon<br/>
+		Ash * Key</h1>
+		<div id=getInvolved>
+			<p style="font-style:italic">The Pokemon inspired
+			ASCII rendered MMORPG</p>
+			<h2>Get Involved</h2>
+			<p>
+				Development is greatly accelerated with the help of people! We need programmers, and artists!
+				Send me a message if you want to help out! <a href="mailto:simon@simonmikulcik.com">simon@simonmikulcik.com</a>
+			</p>
+			<p>
+				If you are shy, feel free to check out the code anyway:
+				<p><a href="https://github.com/smikulcik/pokemon-ash-key-2">Github</a></p>
+				And the Development Organization Strategy:
+				<p><a href="https://trello.com/b/1g1Ng6P8">Trello</a><p>
+		</div>
+	</div>
+	<canvas id="main" width=1000 height=900></canvas>
+</body>
+</html>
